Integrate TitledLinkRule into LinkRule
