+++ markymark/Classes/Flavors/ContentfulFlavor.swift
@@ -22,7 +22,6 @@ open class ContentfulFlavor: Flavor {
     open var defaultRule: Rule = ParagraphRule()
 
     open var inlineRules: [InlineRule] = [
-        TitledLinkRule(),
         LinkRule(),
         BoldRule(),
         ItalicRule(),
